requirements.txt
fasttext-wheel==0.9.2; platform_system == "Windows"
fasttext==0.9.2; platform_system != "Windows"
python-dotenv==1.2.1
Requests==2.32.5
reportlab~=4.4.5